requirements.txt
# --------- pytorch & lightning--------- #
# torch~=2.0.1
lightning==2.0.6
torchmetrics
torchvision

# --------- hydra --------- #
hydra-core~=1.3.2
hydra-colorlog~=1.2.0
hydra-optuna-sweeper~=1.2.0
# omegaconf~=2.3.0

# --------- loggers --------- #
wandb==0.15.8
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
# pyrootutils     # standardizing the project root setup
# pre-commit      # hooks for applying linters on commit
rich==13.5.2      # beautiful text formatting in terminal
pytest==7.4.0
tqdm==4.66.1
# sh            # for running bash commands in some tests (linux/macos only)

# ========= project related ========= #
datasets==2.14.4
accelerate==0.21.0      # If using GPU or on cluster
evaluate~=0.4.0
transformers==4.31.0
jinja2~=3.1.2
# promptsource~=0.2.3
peft~=0.4.0


# torchvision~=0.15.2
# omegaconf~=2.3.0
# packaging~=23.1
# numpy==1.25.2
# deepl~=1.15.0
# psutil~=5.9.5
# setuptools~=67.8.0